refactor file structure.
