+++ src/main/java/com/andersentask/bookshop/book/repositories/BookRepository.java
@@ -1,6 +1,6 @@
-package com.andersentask.bookshop.repositories;
+package com.andersentask.bookshop.book.repositories;
 
-import com.andersentask.bookshop.entities.Book;
+import com.andersentask.bookshop.book.entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
